server-side/requirements.txt
Django
psycopg2-binary
dj-rest-auth
django-allauth
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pillow
channels
daphne
gunicorn
python-dotenv
django-jazzmin
requests
numpy
tensorflow